OT98-92: split class, in this remain getDetailsOrganization's tests
